+++ Assets/02. Scripts/Runtime/Enemies/Model/Properties/SpawnCost.cs
@@ -1,11 +1,11 @@
 ﻿using Runtime.DataFramework.Properties;
 
 namespace Runtime.Enemies.Model.Properties {
-	public interface ISpawnCostProperty : IProperty<int>, ILoadFromConfigProperty {
+	public interface ISpawnCostProperty : IProperty<float>, ILoadFromConfigProperty {
 		
 	}
-	public class SpawnCost : AbstractLoadFromConfigProperty<int>, ISpawnCostProperty {
-		protected override IPropertyDependencyModifier<int> GetDefautModifier() {
+	public class SpawnCost : AbstractLoadFromConfigProperty<float>, ISpawnCostProperty {
+		protected override IPropertyDependencyModifier<float> GetDefautModifier() {
 			return null;
 		}
