+++ Assets/02. Scripts/Runtime/Enemies/SmallEnemies/Bee.cs
@@ -36,7 +36,7 @@ namespace Runtime.Enemies.SmallEnemies
             return level * baseWeight * 2;
         }
 
-        public override int OnGetRealSpawnCost(int level, int rarity, int baseCost) {
+        public override float OnGetRealSpawnCost(int level, int rarity, float baseCost) {
             return level * baseCost * 2;
         }
